FEAT: socket connect/disconnect use
