+++ frontend/app/login/auth/page.tsx
@@ -5,6 +5,7 @@ import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { main } from "@/font/color";
 import { useUser } from "@/context/UserContext";
+import { useAuth } from "@/context/AuthContext";
 
 const modalStyle = {
   position: "absolute" as "absolute",
@@ -23,6 +24,7 @@ const Auth = () => {
   const router = useRouter();
   const [client, setClient] = useState(false);
   const { userDispatch } = useUser();
+  const { authDispatch } = useAuth();
 
   interface Data {
     token: string;
@@ -38,8 +40,8 @@ const Auth = () => {
   const postCode = async (code: string) => {
     // dev original
     // await fetch("http://localhost:4000/login/auth", {
-      // haryu's server
-      await fetch("http://paulryu9309.ddns.net:4000/login/auth", {
+    // haryu's server
+    await fetch("http://paulryu9309.ddns.net:4000/login/auth", {
       method: "POST",
       headers: {
         "Content-type": "application/json",
@@ -51,12 +53,13 @@ const Auth = () => {
     })
       .then(async (res) => {
         if (res.status === 200) {
-          localStorage.setItem("loggedIn", "true");
+          // localStorage.setItem("loggedIn", "true");
           const data: Data = await res.json();
           localStorage.setItem("authorization", data.token); // 서버에서 받은 토큰을 저장
           localStorage.setItem("token", data.jwt);
           localStorage.setItem("intra", data.user.intra);
           localStorage.setItem("idx", data.user.userIdx.toString());
+          authDispatch({ type: "SET_ID", value: data.user.userIdx });
 
           return router.push(`/`);
         }
