+++ frontend/app/page.tsx
@@ -5,6 +5,7 @@ import { useAuth } from "@/context/AuthContext";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { io } from "socket.io-client";
+import { socket } from "./layout";
 
 // dev original
 // export const socket = io("http://localhost:4000/chat", {
@@ -12,17 +13,16 @@ import { io } from "socket.io-client";
 
 const userId =
   typeof window !== "undefined" ? localStorage.getItem("idx") : null;
-// export const socket = io("http://localhost:4000/chat", {
-  // haryu's server
-export const socket = io("http://paulryu9309.ddns.net:4000/chat", {
-  query: { userId: userId },
-});
+// // export const socket = io("http://localhost:4000/chat", {
+//   // haryu's server
+// export const socket = io("http://paulryu9309.ddns.net:4000/chat", {
+//   query: { userId: userId },
+// });
 
 const Page = () => {
   const param = useSearchParams();
   const router = useRouter();
   const [client, setClient] = useState(false);
-  const { authDispatch } = useAuth();
 
   useEffect(() => {
     if (param.get("from") === "game") {
@@ -40,8 +40,7 @@ const Page = () => {
 
   useEffect(() => {
     setClient(true);
-    if (!userId) return;
-    authDispatch({ type: "SET_ID", value: parseInt(userId) });
+    socket.connect();
   }, []);
 
   if (!client) return <></>;
